# cksum: improve performance (Closes: #8573)