--- a/Cargo.toml
+++ b/Cargo.toml
@@ -383,6 +383,7 @@ blake2b_simd = "1.0.2"
 blake3 = "1.5.1"
 sm3 = "0.4.2"
 crc32fast = "1.4.2"
+crc-fast = "1.5.0"
 digest = "0.10.7"
 
 # Fluent dependencies