--- a/src/uucore/Cargo.toml
+++ b/src/uucore/Cargo.toml
@@ -58,6 +58,7 @@ blake2b_simd = { workspace = true, optional = true }
 blake3 = { workspace = true, optional = true }
 sm3 = { workspace = true, optional = true }
 crc32fast = { workspace = true, optional = true }
+crc-fast = { workspace = true, optional = true }
 bigdecimal = { workspace = true, optional = true }
 num-traits = { workspace = true, optional = true }
 selinux = { workspace = true, optional = true }
@@ -151,6 +152,7 @@ sum = [
   "blake3",
   "sm3",
   "crc32fast",
+  "crc-fast",
 ]
 update-control = ["parser"]
 utf8 = []