nice early results:
Still lower but less

```
Benchmark 1: /usr/bin/cksum /tmp/oneline_4G.txt
  Time (mean ± σ):     602.7 ms ±  25.7 ms    [User: 88.0 ms, System: 514.3 ms]
  Range (min … max):   584.5 ms … 620.8 ms    2 runs

Benchmark 2: ./target/release/coreutils cksum /tmp/oneline_4G.txt
  Time (mean ± σ):      1.918 s ±  0.121 s    [User: 0.426 s, System: 1.491 s]
  Range (min … max):    1.832 s …  2.004 s    2 runs

Benchmark 3: ./target/release/coreutils.prev cksum /tmp/oneline_4G.txt
  Time (mean ± σ):      9.502 s ±  0.186 s    [User: 8.893 s, System: 0.608 s]
  Range (min … max):    9.370 s …  9.634 s    2 runs

Summary
  /usr/bin/cksum /tmp/oneline_4G.txt ran
    3.18 ± 0.24 times faster than ./target/release/coreutils cksum /tmp/oneline_4G.txt
   15.77 ± 0.74 times faster than ./target/release/coreutils.prev cksum /tmp/oneline_4G.txt
```